Repackaging: * Combine all source files into a single package * Combine all test files into a single package * Move __main__.py into treescriptify_this package
