+++ test/test_script_writer.py
@@ -2,8 +2,8 @@
 """
 import pytest
 
-from treescriptify.script_writer import generate_script
-from treescriptify.tree_node_data import TreeNodeData
+from treescriptify_this.script_writer import generate_script
+from treescriptify_this.tree_node_data import TreeNodeData
 
 
 def generator_single_dir_node(name: str, depth: int=0):
