+++ treescriptify_this/__init__.py
@@ -1,9 +1,9 @@
 """TreeScriptify Package
 """
-from input.input_data import InputData
-from treescriptify.tree_runner import get_tree_json
-from treescriptify.tree_reader import generate_from_json
-from treescriptify.script_writer import generate_script
+from .input_data import InputData
+from .tree_runner import get_tree_json
+from .tree_reader import generate_from_json
+from .script_writer import generate_script
 
 
 def tsfy(data: InputData) -> str:
