+++ treescriptify_this/argument_parser.py
@@ -2,7 +2,7 @@
 """
 from argparse import ArgumentParser
 
-from input.input_data import InputData
+from treescriptify_this.input_data import InputData
 
 
 def parse_args(arguments: list[str]) -> InputData:
